Move IncrementalInputs to api.execution.incremental
The package `org.gradle.execution` is already taken for some very
non-public stuff.
