+++ buildSrc/subprojects/configuration/src/main/kotlin/org/gradle/gradlebuild/PublicApi.kt
@@ -8,7 +8,6 @@ object PublicApi {
         "org/gradle/caching/**",
         "org/gradle/concurrent/**",
         "org/gradle/deployment/**",
-        "org/gradle/execution/**",
         "org/gradle/external/javadoc/**",
         "org/gradle/ide/**",
         "org/gradle/includedbuild/**",
